requirements.txt
curl_cffi==0.5.10
beautifulsoup4==4.12.3
lxml==5.1.0
rich==13.7.0
tenacity==8.2.3
python-dotenv==1.0.1
